config: configurable cassandra client timeout
Needed this because the large range queries with the dispatcher were
occasionally timing out, especially when the domain is huge. The gocql
default is 600ms and I've set the walker default to 2s, which seems to
cover it. (Note that because this is selecting results in batches and
streaming them, we should not need to keep scaling this up with the size
of the domains. It just seems like our default will need to be larger)