cassandra: added more config items for gocql
The main item I needed was DiscoverHosts, as in production use all
connections being used were for a single machine at a time.

Also, console/model.go was not using cassandra.GetConfig to get a
cassandra config, so it was not using the upped default timeout, thus
the console was more likely to hit timeouts. Now they share, though soon
I'll be doing work to share more stuff.