Added capability to turn on pprof listener
This allows us to turn pprof on, listening on port 6060 so we can
remotely profile it. Enabled only if WALKER_PPROF=1. Used an environment
variable because I assumed that really only developers would want to do
this (though non-devs could still do it), and didn't want to clutter the
config or command line with it since it requires advanced knowledge to
use.